Improve handling of live DOM updates
After recent changes to Goodreads, parts of the DOM get refreshed
just before on just after the pages loads (onload). The previous
commit was intended to handle this but was not reliable in my
testing.

Instead, we watch the parent of the target element (the rating)
with a MutationObserver. This has been 100% successful in my tests.